+++ apps/api/src/domain/models/spatial-feature/errors/index.ts
@@ -1,2 +1,3 @@
 export * from './cannot-replace-traditional-name.error';
 export * from './invalid-geometry-type-for-spatial-feature-error';
+export * from './spatial-feature-must-have-a-name.error';