+++ apps/api/src/domain/models/spatial-feature/errors/spatial-feature-must-have-a-name.error.ts
@@ -1,0 +1,9 @@
+import { InternalError } from '../../../../lib/errors/InternalError';
+
+export class SpatialFeatureMustHaveANameError extends InternalError {
+    constructor() {
+        const msg = `A spatial feature must have at least one of: traditional name, contemporary name`;
+
+        super(msg);
+    }
+}