+++ apps/api/src/domain/models/spatial-feature/interfaces/geometric-feature.interface.ts
@@ -1,8 +1,7 @@
 import { GeometricCoordinatesUnion } from '../types/Coordinates/GeometricCoordinatesUnion';
-import { GeometricFeatureType } from '../types/GeometricFeatureType';
 
 export interface IGeometricFeature<
-    TGeometricFeatureType extends GeometricFeatureType = GeometricFeatureType,
+    TGeometricFeatureType extends string = string,
     UCoordinate extends GeometricCoordinatesUnion = GeometricCoordinatesUnion
 > {
     type: TGeometricFeatureType;