+++ apps/api/src/domain/models/spatial-feature/point/commands/create-point.command-handler.ts
@@ -1,5 +1,12 @@
-import { AggregateType, GeometricFeatureType, ResourceType } from '@coscrad/api-interfaces';
+import {
+    AggregateType,
+    GeometricFeatureType,
+    MultilingualTextItemRole,
+} from '@coscrad/api-interfaces';
 import { CommandHandler } from '@coscrad/commands';
+import { isNonEmptyObject } from '@coscrad/validation-constraints';
+import { buildMultilingualTextWithSingleItem } from '../../../../../domain/common/build-multilingual-text-with-single-item';
+import { MultilingualTextItem } from '../../../../../domain/common/entities/multilingual-text';
 import { InternalError } from '../../../../../lib/errors/InternalError';
 import { ResultOrError } from '../../../../../types/ResultOrError';
 import { Valid } from '../../../../domainModelValidators/Valid';
@@ -8,15 +15,16 @@ import { InMemorySnapshot } from '../../../../types/ResourceType';
 import { BaseCreateCommandHandler } from '../../../shared/command-handlers/base-create-command-handler';
 import { BaseEvent } from '../../../shared/events/base-event.entity';
 import { EventRecordMetadata } from '../../../shared/events/types/EventRecordMetadata';
-import { validAggregateOrThrow } from '../../../shared/functional';
 import { Point } from '../entities/point.entity';
 import { CreatePoint } from './create-point.command';
-import { PointCreated } from './point-created.event';
+import { PointCreated, PointCreatedPayload } from './point-created.event';
 
 @CommandHandler(CreatePoint)
 export class CreatePointCommandHandler extends BaseCreateCommandHandler<Point> {
     protected createNewInstance({
         aggregateCompositeIdentifier: { id },
+        traditionalName,
+        contemporaryName,
         lattitude,
         longitude,
         // TODO remove this
@@ -24,7 +32,21 @@ export class CreatePointCommandHandler extends BaseCreateCommandHandler<Point> {
         description,
         imageUrl,
     }: CreatePoint): ResultOrError<Point> {
-        return new Point({
+        const tradionalNameToUse = isNonEmptyObject(traditionalName)
+            ? buildMultilingualTextWithSingleItem(
+                  traditionalName.text,
+                  traditionalName.languageCode
+              )
+            : null;
+
+        const contemporaryNameToUse = isNonEmptyObject(contemporaryName)
+            ? buildMultilingualTextWithSingleItem(
+                  contemporaryName.text,
+                  contemporaryName.languageCode
+              )
+            : null;
+
+        const instance = new Point({
             type: AggregateType.spatialFeature,
             id,
             geometry: {
@@ -32,34 +54,55 @@ export class CreatePointCommandHandler extends BaseCreateCommandHandler<Point> {
                 coordinates: [lattitude, longitude],
             },
             properties: {
-                // this too
-                // name,
+                traditionalName: tradionalNameToUse,
+                contemporaryName: contemporaryNameToUse,
                 description,
                 imageUrl,
             },
             // You must run a `PUBLISH_RESOURCE` command to publish this point
             published: false,
         });
+
+        const validationResult = instance.validateInvariants();
+
+        if (validationResult !== Valid) {
+            return new InternalError(`Failed to create point/${id}.`, [validationResult]);
+        }
+
+        return instance;
     }
 
     protected async fetchRequiredExternalState(_?: CreatePoint): Promise<InMemorySnapshot> {
-        const allSpatialFeatures = (
-            await this.repositoryProvider.forResource(ResourceType.spatialFeature).fetchMany()
-        ).filter(validAggregateOrThrow);
-
-        return new DeluxeInMemoryStore({
-            [AggregateType.spatialFeature]: allSpatialFeatures,
-        }).fetchFullSnapshotInLegacyFormat();
+        return new DeluxeInMemoryStore({}).fetchFullSnapshotInLegacyFormat();
     }
 
     protected validateExternalState(
-        externalState: InMemorySnapshot,
-        point: Point
+        _externalState: InMemorySnapshot,
+        _point: Point
     ): InternalError | Valid {
-        return point.validateExternalState(externalState);
+        /**
+         * We do not require place names to be unique.
+         */
+        return Valid;
     }
 
     protected buildEvent(command: CreatePoint, eventMeta: EventRecordMetadata): BaseEvent {
-        return new PointCreated(command, eventMeta);
+        const eventPayload: PointCreatedPayload = {
+            ...command,
+            traditionalName: command.traditionalName
+                ? new MultilingualTextItem({
+                      ...command.traditionalName,
+                      role: MultilingualTextItemRole.original,
+                  })
+                : undefined,
+            contemporaryName: command.traditionalName
+                ? new MultilingualTextItem({
+                      ...command.contemporaryName,
+                      role: MultilingualTextItemRole.original,
+                  })
+                : undefined,
+        };
+
+        return new PointCreated(eventPayload, eventMeta);
     }
 }