+++ apps/api/src/domain/models/spatial-feature/point/commands/create-point.command.integration.spec.ts
@@ -1,11 +1,12 @@
+import { LanguageCode } from '@coscrad/api-interfaces';
 import { CommandHandlerService } from '@coscrad/commands';
 import { INestApplication } from '@nestjs/common';
 import setUpIntegrationTest from '../../../../../app/controllers/__tests__/setUpIntegrationTest';
 import { clonePlainObjectWithOverrides } from '../../../../../lib/utilities/clonePlainObjectWithOverrides';
 import TestRepositoryProvider from '../../../../../persistence/repositories/__tests__/TestRepositoryProvider';
 import generateDatabaseNameForTestSuite from '../../../../../persistence/repositories/__tests__/generateDatabaseNameForTestSuite';
+import { buildTestInstance } from '../../../../../test-data/utilities';
 import { getCommandFsaForTest } from '../../../../__tests__/utilities/getCommandFsaForTest';
-import { getValidSpatialFeatureInstanceForTest } from '../../../../__tests__/utilities/getValidSpatialFeatureInstanceForTest';
 import { IIdManager } from '../../../../interfaces/id-manager.interface';
 import { AggregateId } from '../../../../types/AggregateId';
 import { AggregateType } from '../../../../types/AggregateType';
@@ -18,16 +19,20 @@ import { generateCommandFuzzTestCases } from '../../../__tests__/command-helpers
 import { CommandAssertionDependencies } from '../../../__tests__/command-helpers/types/CommandAssertionDependencies';
 import buildDummyUuid from '../../../__tests__/utilities/buildDummyUuid';
 import { dummySystemUserId } from '../../../__tests__/utilities/dummySystemUserId';
-import { GeometricFeatureType } from '../../types/GeometricFeatureType';
 import { CreatePoint } from './create-point.command';
 
 const commandType = `CREATE_POINT`;
 
 const pointName = 'Sunny Park Point';
 
+const traditionalName = {
+    text: pointName,
+    languageCode: LanguageCode.Chilcotin,
+};
+
 const dummyFsa = getCommandFsaForTest<CreatePoint>(commandType, {
-    aggregateCompositeIdentifier: { id: buildDummyUuid(55) },
-    name: pointName,
+    aggregateCompositeIdentifier: { id: buildDummyUuid(101) },
+    traditionalName,
 });
 
 const commandFsaFactory = new DummyCommandFsaFactory((id) => {
@@ -78,35 +83,16 @@ describe(commandType, () => {
         it(`should succeed with the expected updates to the database`, async () => {
             await assertCreateCommandSuccess(assertionHelperDependencies, {
                 systemUserId: dummySystemUserId,
-                initialState: new DeluxeInMemoryStore({}).fetchFullSnapshotInLegacyFormat(),
+                seedInitialState: async () => {
+                    Promise.resolve();
+                },
                 buildValidCommandFSA: (id: AggregateType) => commandFsaFactory.build(id),
             });
         });
     });
 
     describe(`when the command is invalid`, () => {
-        describe(`when there is already another spatial feature with the given name`, () => {
-            Object.values(GeometricFeatureType).forEach((featureType) => {
-                describe(`with geometry type: ${featureType}`, () => {
-                    it(`should fail with the expected errors`, async () => {
-                        await assertCreateCommandError(assertionHelperDependencies, {
-                            systemUserId: dummySystemUserId,
-                            initialState: new DeluxeInMemoryStore({
-                                [AggregateType.spatialFeature]: [
-                                    getValidSpatialFeatureInstanceForTest(featureType).clone({
-                                        properties: {
-                                            name: pointName,
-                                            description: 'My name is already taken!',
-                                        },
-                                    }),
-                                ],
-                            }).fetchFullSnapshotInLegacyFormat(),
-                            buildCommandFSA: (id: AggregateId) => commandFsaFactory.build(id),
-                        });
-                    });
-                });
-            });
-        });
+        //  note that we do not want to constrain only one place to have a given name as this will surely happen across different communities and families
 
         describe('when the id has not been generated via our system', () => {
             it('should return the expected error', async () => {
@@ -152,6 +138,72 @@ describe(commandType, () => {
                     });
                 }
             );
+        });
+
+        describe(`when the geometry is invalid`, () => {
+            it(`should have a test`, async () => {
+                const invalidLattitude = 9000;
+
+                // note that -Infinity would be caught at the schema level- fuzz test this!
+                const invalidLongitude = -300;
+
+                await assertCreateCommandError(assertionHelperDependencies, {
+                    systemUserId: dummySystemUserId,
+                    seedInitialState: async () => {
+                        Promise.resolve();
+                    },
+                    buildCommandFSA: (pointId) => ({
+                        type: commandType,
+                        payload: buildTestInstance(CreatePoint, {
+                            aggregateCompositeIdentifier: {
+                                id: pointId,
+                            },
+                            contemporaryName: {
+                                text: 'blue rock',
+                                languageCode: LanguageCode.English,
+                            },
+                            lattitude: invalidLattitude,
+                            longitude: invalidLongitude,
+                        }),
+                    }),
+                    checkError: (error) => {
+                        const msg = error.toString();
+
+                        expect(msg).toContain(invalidLattitude.toString());
+                        expect(msg).toContain(invalidLongitude.toString());
+                    },
+                });
+            });
+        });
+
+        describe(`when no name is provided`, () => {
+            it(`should fail with the expected error`, async () => {
+                await assertCreateCommandError(assertionHelperDependencies, {
+                    systemUserId: dummySystemUserId,
+                    seedInitialState: async () => {
+                        Promise.resolve();
+                    },
+                    buildCommandFSA: (pointId) => ({
+                        type: commandType,
+                        payload: buildTestInstance(CreatePoint, {
+                            aggregateCompositeIdentifier: {
+                                id: pointId,
+                            },
+                            traditionalName: null,
+                            contemporaryName: null,
+                        }),
+                    }),
+                    checkError: (error, pointId) => {
+                        const msg = error.toString();
+
+                        expect(msg).toContain(pointId);
+
+                        expect(msg).toContain(
+                            `must have at least one of: traditional name, contemporary name`
+                        );
+                    },
+                });
+            });
         });
     });
 });