+++ apps/api/src/test-data/buildSpatialFeatureTestData/buildLineTestData.ts
@@ -1,3 +1,5 @@
+import { LanguageCode } from '@coscrad/api-interfaces';
+import { buildMultilingualTextWithSingleItem } from '../../domain/common/build-multilingual-text-with-single-item';
 import { Line } from '../../domain/models/spatial-feature/line/entities/line.entity';
 import { GeometricFeatureType } from '../../domain/models/spatial-feature/types/GeometricFeatureType';
 import { ResourceType } from '../../domain/types/ResourceType';
@@ -21,7 +23,11 @@ const dtos: DTO<Line>[] = [
             ],
         },
         properties: {
-            name: 'Windy Path',
+            traditionalName: buildMultilingualTextWithSingleItem(
+                'Windy Path',
+                LanguageCode.Chilcotin
+            ),
+            contemporaryName: buildMultilingualTextWithSingleItem('It means windy path'),
             description: 'One amazing hike!',
             imageUrl: 'https://coscrad.org/wp-content/uploads/2023/05/wind-energy-2029621_1280.png',
         },