+++ apps/api/src/test-data/buildSpatialFeatureTestData/buildPolygonTestData.ts
@@ -1,3 +1,5 @@
+import { LanguageCode } from '@coscrad/api-interfaces';
+import { buildMultilingualTextWithSingleItem } from '../../domain/common/build-multilingual-text-with-single-item';
 import { IGeometricFeature } from '../../domain/models/spatial-feature/interfaces/geometric-feature.interface';
 import { Polygon } from '../../domain/models/spatial-feature/polygon/entities/polygon.entity';
 import { PolygonCoordinates } from '../../domain/models/spatial-feature/types/Coordinates/PolygonCoordinates';
@@ -30,7 +32,14 @@ export default (): Polygon[] =>
                 id: `${index + 300}`,
                 published: true,
                 properties: {
-                    name: `Point with ID: ${`${index + 300}`} Name`,
+                    traditionalName: buildMultilingualTextWithSingleItem(
+                        `Traditional Name of Point with ID: ${`${index + 300}`} Name`,
+                        LanguageCode.Chilcotin
+                    ),
+                    contemporaryName: buildMultilingualTextWithSingleItem(
+                        `Contemporary Name of Point with ID: ${`${index + 300}`} Name`,
+                        LanguageCode.Chilcotin
+                    ),
                     description: `Description for polygon ${`${index + 300}`}`,
                     imageUrl: 'https://coscrad.org/wp-content/uploads/2023/05/map-1272165_640.png',
                 },