+++ libs/api-interfaces/src/lib/aggregate-views/view-models/resources/spatial-feature/spatial-feature.view-model.interface.ts
@@ -1,8 +1,9 @@
 import { IBaseViewModel } from '../../base.view-model.interface';
+import { IMultilingualText } from '../common';
 import { ISpatialFeatureGeometry } from './spatial-feature-geometry.interface';
 import { ISpatialFeatureProperties } from './spatial-feature-properties.interface';
 
 export interface ISpatialFeatureViewModel<T = unknown> extends IBaseViewModel {
     geometry: ISpatialFeatureGeometry<T>;
-    properties: ISpatialFeatureProperties;
+    properties: ISpatialFeatureProperties & { name: IMultilingualText };
 }